refactor: 抽取clientLoadery和routeProps公用逻辑
